requirements.txt
# Core
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.29.0

# HTTP & WebSocket
httpx>=0.27.0
websockets>=12.0

# Auth
PyJWT>=2.8.0

# Scheduler
apscheduler>=3.10.0

# Env
python-dotenv>=1.0.0

# YAML
pyyaml>=6.0.0

# Dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
